Added sample config file
File named "konfig" is now located in the docs subdirectory. In order to
use sirbot for the time being, move the file to the sirbot\config\sirbot
directory and rename to "config". Open the file with notepad or other
plain text editor and towards the end change USERNAME_HERE to the
username you wish to log in with, TOKEN to the associated oauth token
(leave off 'oauth:'), and CHANNEL_HERE to the name of the channel you
wish to join (no leading '#' is necessary) and then you are ready to run
the application which will for the time being automatically join the
channel entered. All chat text is raw currently.