A lot of changes happened here.
Now importing choice from random. Also added a time getter in the main
class. Now keeping track of the current time and when things are run.
Also altered the way in which sub groups were stored and changed the
deletion algorithm to match. Also began implementing the function for
actually responding to commands. Also formalized the way in which
messages would be passed to the ai file.