Single word commands are now functional.
Also now keeping track of the bot's name. Also returning things to be
printed as a list where the first item is 1 if it is to be printed/sent
to chat. Made it so a remainder or tempval only command cannot exist.
Updated the test to test the new running of commands.